fix: separate multiaddr from network
